+++ Arduino/Test/PanelTest/_DGraphRoling.ino
@@ -13,12 +13,12 @@ byte green = 0;
      // initialize panel layout
       Panel.send(ApplicationName,F("PanelTest - RolingGraph")); // set the application name
       Panel.send(PanelColor,"$PINK"); // set the application name
-      Panel.send(DynamicDisplay,100);     // enable dynamic display request
+      Panel.send(DynamicDisplay,(int16_t) 100);     // enable dynamic display request
       Panel.send(Button_4,  F("next\npanel"));
       Panel.send(Button_5,  F("info"));
       Panel.send(Button_6,  F("moni\ntor"));
       Panel.send(Button_7,  F("graph"));
-      Panel.send(GraphGrid, 15);
+      Panel.send(GraphGrid, (int16_t)15);
       Panel.send(GraphCaption_1, F("Caption_1"));
       Panel.send(GraphCaption_2, F("Caption_2"));
       Panel.send(GraphButton_1, F("GrBttn 1"));
